readd scale, but enlarge smaller images

+++ ppmclibs/tinycv.h
@@ -22,5 +22,5 @@ void image_threshold(Image *s, int level);
 std::vector<float> image_avgcolor(Image *s);
 bool image_differ(Image *a, Image *b, unsigned char maxdiff);
 
-Image *image_scale(Image *a, long width, long height);
+Image *image_scale(Image *a, int width, int height);
 double image_similarity(Image *a, Image*b);
